GitLab CI: allow forcing to run a pipeline by setting $CI_FORCE_RUN
This is useful for example when someone without commit access created a MR for
a branch in their own repository. In this case, we don't get Jenkins nor GitLab
CI results, so usually we will push a copy of the branch to the main repo.
This will give us Jenkins results, but not GitLab CI.

With this change, one can manually trigger a GitLab CI pipeline run,
setting the $CI_FORCE_RUN pipeline variable.